Serve actual file
Before this commit we just served the same static html page for every
path. Now we actually use the path to get the correct resource. To make
this work properly we also need to set the `Content-Type` header, so the
store now stores a `Resource` which can be used to derive the correct
mime-type.