Allow processors to return extra resources when finished
Introduce a `flush` method which returns a `Vec` of extra resources and
their path on the website. Returning the path is a bit of a hack -
currently `Resource` paths are all full filesystem paths, which is a bit
tricky when we just want to state what URL to find the resource at.