`Resource`s now contain their url path
Before this commit `Resource`s were a bit strange because they contained
filesystem paths which were then mapped to url paths by removing the
base filepath we built the store path. This meant for
`ResourceProcessor::flush` we needed to give the url path in a tuple.

To fix this `Resource`s now have a field which instructs how to build
the url path. It can either be from the original path, from a different
filesystem path or you can just specify the url path directly.