Add SSE route for when we have rebuilt
Server Side Events are a way of sending events from the server to the
browser. We will use them to tell the web page to reload if in
development mode. We enable this functionality by creating a broadcast
channel which has a message sent to it when a rebuild is finished.

Note that we only enable this in development mode which is turned on by
passing the `-d` flag.